+++ tools/test_v7_3.js
@@ -1,3 +1,0 @@
-'use strict';
-// Compatibilidade: a suíte vigente do projeto está em test_v7_3_1.js.
-require('./test_v7_3_1.js');